change some "defaultValue"s
Key "theme": I also use the dark version of F-Droid but I think the most users would prefer the light version. This is also standard in the
most used apps on Android, I think, see GApps, Flym, K-9, Telegram, ...

Key "localRepoBonjour": I never used this function and I think it's better, if the user manually activates it when he wants to use it.

Key "localRepoHttps": I see no reason why it should not use https over http, it's better of security view, I think.

Key "cacheDownloaded": The most smartphones today have enough storage available to allow to enable this function. This would decrease the
traffic on both sides, as the user does not have to download a app and the F-Droid server does not have to send it multiple
times.