CI: 4.4 says it's ready before it actually is
Its boot animation is "stopped" but on my laptop the CPU is still at 100%,
which means it's probably still working and not ready yet. The tests fail if
you run them right after wait-for-emulator is done. If you wait a few seconds
though, they do work since the CPU load drops.

Until we can figure out a better solution, have a generous 30-second sleep
after it tells us it's ready.